+++ packages/author-head/author-head.js
@@ -28,7 +28,8 @@ const styles = StyleSheet.create({
     height: 100,
     borderColor: "#FFF",
     borderRadius: 50,
-    borderWidth: 5
+    borderWidth: 5,
+    overflow: "hidden"
   },
   name: {
     fontFamily: "TimesModern-Bold",
@@ -39,7 +40,13 @@ const styles = StyleSheet.create({
   title: {
     fontFamily: "TimesDigital-RegularSC",
     fontSize: 15,
-    color: "#696969"
+    color: "#696969",
+    ...Platform.select({
+      web: {
+        WebkitFontSmoothing: "antialiased",
+        MozOsxFontSmoothing: "grayscale"
+      }
+    })
   },
   twitter: {
     fontSize: 18,
@@ -56,7 +63,13 @@ const styles = StyleSheet.create({
     lineHeight: 26,
     color: "#333",
     maxWidth: "88%",
-    paddingBottom: 32
+    paddingBottom: 32,
+    ...Platform.select({
+      web: {
+        WebkitFontSmoothing: "antialiased",
+        MozOsxFontSmoothing: "grayscale"
+      }
+    })
   },
   wrapper: {
     alignItems: "center",
@@ -70,14 +83,19 @@ const AuthorHead = props => {
 
   const imageComponent = uri ? (
     <View style={styles.photoContainer}>
-      <Image source={{ uri }} style={styles.roundImage} />
+      <Image uri={uri} style={styles.roundImage} aspectRatio={1 / 1} />
     </View>
   ) : null;
 
   return (
     <View style={styles.wrapper} pointerEvents="box-none">
       <View accessibilityRole="banner" style={styles.container}>
-        <Text accessibilityRole="heading" aria-level="1" style={styles.name}>
+        <Text
+          testID="author-name"
+          accessibilityRole="heading"
+          aria-level="1"
+          style={styles.name}
+        >
           {name}
         </Text>
         <Text accessibilityRole="heading" aria-level="2" style={styles.title}>