+++ packages/brightcove-video/__tests__/brightcove-player.web.test.js
@@ -326,35 +326,6 @@ describe("brightcove-video web component", () => {
           }, 50);
         });
 
-        it("will emit a 'finish' event if time at pause is greater than or equal to video duration", done => {
-          dummyPlayer.currentTime = () => 11;
-          dummyPlayer.duration = () => 11;
-
-          const component = (
-            <BrightcoveVideo
-              accountId="57838016001"
-              videoId="[X]"
-              onFinish={() => {
-                done();
-              }}
-            />
-          );
-
-          ReactDOM.render(component, reactWrapper);
-
-          setTimeout(() => {
-            dummyScript.onload();
-          }, 50);
-
-          setTimeout(() => {
-            evtReg.durationchange();
-          }, 100);
-
-          setTimeout(() => {
-            evtReg.pause();
-          }, 150);
-        });
-
         it("will emit a 'progress' event", done => {
           dummyPlayer.currentTime = () => 0.1;
 
@@ -402,6 +373,28 @@ describe("brightcove-video web component", () => {
 
             setTimeout(() => {
               evtReg.durationchange();
+            }, 50);
+          }, 50);
+        });
+
+        it("will emit a 'finish' event", done => {
+          const component = (
+            <BrightcoveVideo
+              accountId="57838016001"
+              videoId="[X]"
+              onFinish={() => {
+                done();
+              }}
+            />
+          );
+
+          ReactDOM.render(component, reactWrapper);
+
+          setTimeout(() => {
+            dummyScript.onload();
+
+            setTimeout(() => {
+              evtReg.ended();
             }, 50);
           }, 50);
         });