+++ packages/pagination/pagination.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { StyleSheet, Text, View } from "react-native";
+import { StyleSheet, Text, View, Platform } from "react-native";
 import PropTypes from "prop-types";
 import { TextLink } from "@times-components/link";
 import withPageState from "./pagination-wrapper";
@@ -35,7 +35,13 @@ const styles = StyleSheet.create({
   label: {
     color: "#696969",
     fontFamily: "GillSansMTStd-Medium",
-    fontSize: 15
+    fontSize: 15,
+    ...Platform.select({
+      web: {
+        WebkitFontSmoothing: "antialiased",
+        MozOsxFontSmoothing: "grayscale"
+      }
+    })
   },
   message: {
     justifyContent: "center"
@@ -110,7 +116,7 @@ class Pagination extends React.Component {
           onPress={(...params) => onPrev(page - 1, ...params)}
           url={generatePageLink(page - 1)}
         >
-          {"< Previous page"}
+          {"< Previous"}
         </TextLink>
       ) : null;
 
@@ -121,7 +127,7 @@ class Pagination extends React.Component {
           onPress={(...params) => onNext(page + 1, ...params)}
           url={generatePageLink(page + 1)}
         >
-          {"Next page >"}
+          {"Next >"}
         </TextLink>
       ) : null;