+++ packages/author-head/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.6.12"></a>
+## [0.6.12](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.11...@times-components/author-head@0.6.12) (2017-10-05)
+
+
+
+
+<a name="0.6.11"></a>
+## [0.6.11](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.10...@times-components/author-head@0.6.11) (2017-10-04)
+
+
+
+
+<a name="0.6.10"></a>
+## [0.6.10](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.9...@times-components/author-head@0.6.10) (2017-10-03)
+
+
+
+
+<a name="0.6.9"></a>
+## [0.6.9](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.8...@times-components/author-head@0.6.9) (2017-10-03)
+
+
+### Bug Fixes
+
+* web font smoothing ([c3aa351](https://github.com/newsuk/times-components/commit/c3aa351))
+
+
+
+
 <a name="0.6.8"></a>
 ## [0.6.8](https://github.com/newsuk/times-components/compare/@times-components/author-head@0.6.7...@times-components/author-head@0.6.8) (2017-09-28)