+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.4.5"></a>
+## [0.4.5](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.4.4...@times-components/error-view@0.4.5) (2017-10-05)
+
+
+
+
+<a name="0.4.4"></a>
+## [0.4.4](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.4.3...@times-components/error-view@0.4.4) (2017-10-04)
+
+
+
+
+<a name="0.4.3"></a>
+## [0.4.3](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.4.2...@times-components/error-view@0.4.3) (2017-10-04)
+
+
+
+
+<a name="0.4.2"></a>
+## [0.4.2](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.4.1...@times-components/error-view@0.4.2) (2017-10-03)
+
+
+
+
 <a name="0.4.1"></a>
 ## [0.4.1](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.4.0...@times-components/error-view@0.4.1) (2017-09-25)